Refatorando a Calculadora.
