+++ src/Porter2023/Interfaces/ICalculadoraExpressao.cs
@@ -1,0 +1,7 @@
+﻿namespace Porter2023.Interfaces
+{
+    public interface ICalculadoraExpressao
+    {
+        string Calcular(string expressao);
+    }
+}
