+++ src/Porter2023/Interfaces/ISomadorArray.cs
@@ -1,0 +1,7 @@
+﻿namespace Porter2023.Interfaces
+{
+    public interface ISomadorArray
+    {
+        long Somar(int[] numeros);
+    }
+}
